python-ocr/requirements.txt
flask==3.0.0
Werkzeug==3.0.1
easyocr==1.7.1
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
Pillow==10.4.0
pdf2image==1.17.0
python-dotenv==1.0.1

